[Refactor] Only run the error count metrics after midnight
